moves misc signal component funcs to utils module
