+++ moped-editor/src/utils/signalComponentHelpers.js
@@ -1,3 +1,4 @@
+import { useEffect } from "react";
 import { v4 as uuidv4 } from "uuid";
 
 /**
@@ -33,4 +34,51 @@ import { v4 as uuidv4 } from "uuid";
     featureCollection.features.push(feature);
   }
   return featureCollection;
+};
+
+export const useInitialSignalComponentValue = (editFeatureCollection, setSignal) => {
+  /*
+  / initializes the selected signal value - handles case of editing existing component
+  / tests for signal_id prop to ensure we're not handing a non-signal component (which happens
+  / e.g. when an existing component's type is changed)
+  */
+  useEffect(() => {
+    if (
+      !editFeatureCollection ||
+      editFeatureCollection.features.length === 0 ||
+      !editFeatureCollection.features[0].properties.signal_id
+    ) {
+      setSignal(null);
+      return;
+    } else if (editFeatureCollection.features.length > 1) {
+      /*
+      / If a non-signal component is edited, all previously-defined feature geometries
+      / will be dropped.
+      */
+      console.warn(
+        "Found signal component with multiple feature geometries. All but one feature will be removed."
+      );
+    }
+    setSignal(editFeatureCollection.features[0]);
+    // only fire on init
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+};
+
+/*
+/ Hook which updates the component editor state based on the selected signal
+*/
+export const useSignalChangeEffect = (
+  signal,
+  setSelectedComponentSubtype,
+  setEditFeatureCollection
+) => {
+  useEffect(() => {
+    const signalSubtype = signal
+      ? signal.properties.signal_type.toLowerCase()
+      : "";
+    const featureCollection = signalToFeatureCollection(signal);
+    setSelectedComponentSubtype(signalSubtype);
+    setEditFeatureCollection(featureCollection);
+  }, [signal, setSelectedComponentSubtype, setEditFeatureCollection]);
 };
